+++ front-end/src/ManagerPage.jsx
